Python3, cleanup and removal of ini file
I don't see many cases where you join a channel, but don't want to see
notifications about it. As that's the case, why would I have to manually
enter the channels I use a second time in an ini file in order to get
notifications. Simply show all messages in a channel.

Besides, I switched to python3, which only required updating the
print's, and I cleaned the code a bit, and used visual indentation.